backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
drf-yasg==1.21.7
python-dotenv==1.0.1
setuptools>=65.0.0
